refactor: split routes
