+++ main.py
@@ -241,166 +241,6 @@ async def home(request: Request, user=Depends(require_login)):
     return await render(request, "index.html", {"resumes": resumes}, page_title="Home")
 
 
-@app.get("/chat", response_class=HTMLResponse)
-async def chat_interface(request: Request, user=Depends(require_login)):
-    # sidebar list (may be empty if Mongo down)
-    candidates = [
-        {"id": r["resume_id"], "name": r["name"]}
-        for r in await resumes_all()
-    ]
-
-    session_user = await get_current_user(request.cookies.get(COOKIE_NAME))
-    user_id = session_user["username"] if session_user else "anon"
-    doc      = await chat_find_one({"user_id": user_id}) or {}
-    history  = doc.get("messages", [])
-
-    if not isinstance(history, list):    # ⬅️ guard against bad data
-        history = []
-
-    last_proj = doc.get("last_project", {})   # (keep if you still need it)
-
-    safe_history = [
-        {"role": m.get("role"), "content": sanitize_markdown(m.get("content", ""))}
-        for m in history if isinstance(m, dict)
-    ]
-
-    return await render(
-        request,
-        "chat.html",
-        {"history": safe_history, "candidates": candidates},
-        page_title="Chat",
-    )
-
-
-# ──────────────────────────────────────────────────────────────────────────
-# /chat  – text follow-ups (no file upload)
-# ──────────────────────────────────────────────────────────────────────────
-@app.post("/chat", response_class=JSONResponse)
-async def chat(request: Request,
-               chat_data: ChatRequest = Body(...),
-               user=Depends(require_login)):
-    user_text    = chat_data.text.strip()
-    selected_ids = chat_data.candidate_ids        # list[str]
-    session_user = await get_current_user(request.cookies.get(COOKIE_NAME))
-    user_id      = session_user["username"] if session_user else "anon"
-
-    # 1️⃣ fetch previous convo & project context
-    doc        = await chat_find_one({"user_id": user_id}) or {}
-    history    = doc.get("messages", [])
-    last_proj  = doc.get("last_project", {})               # may be {}
-
-    # 2️⃣ guarantee history is a list (avoid slice KeyError)
-    if not isinstance(history, list):
-        history = []
-
-    # 3️⃣ build résumé snippets to inject (filtered by sidebar checkboxes)
-    snippets = []
-    for c in last_proj.get("candidates", []):
-        if not selected_ids or c["name"] in selected_ids:
-            snippets.append(
-                f"**{c['name']}** ({c['fit']} %)\n{c['text'][:600]}…"
-            )
-
-    # 4️⃣ system-level context blocks
-    system_blocks = []
-    if desc := last_proj.get("description"):
-        system_blocks.append(f"### Project\n{desc}")
-    if snippets:
-        system_blocks.append("### Candidate résumés\n" + "\n\n".join(snippets))
-
-    # 5️⃣ assemble message list for GPT
-    messages = [
-        {
-            "role": "system",
-            "content": (
-                "You are a recruitment assistant who can answer follow-up "
-                "questions about the project and the candidate résumés provided."
-            ),
-        }
-    ]
-    if system_blocks:
-        messages.append({"role": "system", "content": "\n\n".join(system_blocks)})
-
-    # last 6 legitimate turns (skip empty / malformed entries)
-    for turn in history[-6:]:
-        if isinstance(turn, dict) and turn.get("content", "").strip():
-            messages.append({"role": turn["role"], "content": turn["content"]})
-
-    # current user turn
-    messages.append({"role": "user", "content": user_text})
-
-    # 6️⃣ call OpenAI
-    resp = openai.chat.completions.create(
-        model="gpt-3.5-turbo",
-        messages=messages,
-        temperature=0.4,
-        max_tokens=600,
-    )
-    assistant_reply = resp.choices[0].message.content.strip()
-    safe_reply = sanitize_markdown(assistant_reply)
-
-    # 7️⃣ persist updated history
-    history.extend(
-        [
-            {"role": "user", "content": user_text},
-            {"role": "assistant", "content": assistant_reply},
-        ]
-    )
-    await chat_upsert(user_id, {"messages": history})
-
-    return {"reply": safe_reply}
-
-
-# ═════════════════════════════════════════════════════════════════════════════
-# /chat_action – simple buttons (top‑5, python expert)
-# ═════════════════════════════════════════════════════════════════════════════
-
-@app.post("/chat_action", response_class=HTMLResponse)
-async def chat_action(
-    request: Request,
-    user=Depends(require_login),
-    action: str = Form(...),
-    candidate_ids: List[str] = Form([]),
-):
-    session_user = await get_current_user(request.cookies.get(COOKIE_NAME))
-    user_id = session_user["username"] if session_user else "anon"
-    if action == "top5":
-        user_instr = "Pick the top 5 candidates and explain briefly why."
-    elif action == "python_expert":
-        user_instr = "Which candidate has the most Python experience?"
-    else:
-        user_instr = "Sorry, I didn’t understand that request."
-
-    # candidate set
-    if candidate_ids:
-        matches = await resumes_by_ids(candidate_ids)
-    else:
-        matches = search_best_resumes("(the project text…)", top_k=10, namespace="resumes")[:5]
-
-    block = "".join(
-        f"---\nName: {m.metadata['name']}\nRésumé:\n{m.metadata['text']}\n\n" for m in matches
-    )
-
-    prompt = (
-        "You are a helpful recruitment assistant.\n\n" + user_instr + "\n\nHere are the candidates:\n" + block
-    )
-
-    reply = openai.chat.completions.create(
-        model="gpt-3.5-turbo",
-        messages=[
-            {"role": "system", "content": "You are a helpful recruitment assistant."},
-            {"role": "user", "content": prompt},
-        ],
-    ).choices[0].message.content
-    safe_reply = sanitize_markdown(reply)
-
-    candidates = [{"id": r["resume_id"], "name": r["name"]} for r in await resumes_all()]
-    return await render(
-        request,
-        "chat.html",
-        {"reply": safe_reply, "candidates": candidates},
-        page_title="Chat",
-    )
 
 # ═════════════════════════════════════════════════════════════════════════════
 # upload résumé – saves to Pinecone + Mongo (if available)
@@ -901,199 +741,8 @@ async def delete_resume_route(id: str = Form(...)):
 
     return RedirectResponse("/resumes", status_code=303)
 
+from routes.users import router as users_router, login_form, login_post, logout, user_admin, create_user_admin, delete_user_admin, edit_user_form, update_user_admin, reset_user_password, profile, change_password
+from routes.chat import router as chat_router, chat_interface, chat, chat_action, chat_followup
 
-@app.post("/chat_followup", response_class=JSONResponse)
-async def chat_followup(request: Request, user=Depends(require_login)):
-    payload        = await request.json()
-    user_text      = payload.get("message", "").strip()
-    session_user = await get_current_user(request.cookies.get(COOKIE_NAME))
-    user_id        = session_user["username"] if session_user else "anon"
-
-    doc = await chats.find_one({"user_id": user_id}) or {}
-    history  = doc.get("messages", [])
-    project  = doc.get("last_project")
-
-    # ── 1. if no project context fall back to generic model ────────────────
-    if not project:
-        history.append({"role":"user","content":user_text})
-        messages = (
-            [{"role":"system","content":"You are a helpful recruitment assistant."}]
-            + history
-        )
-        reply = openai.chat.completions.create(
-            model="gpt-3.5-turbo",
-            messages=messages
-        ).choices[0].message.content
-        history.append({"role":"assistant","content":reply})
-        await chats.update_one({"user_id":user_id},{"$set":{"messages":history}},upsert=True)
-        return {"reply": reply}
-
-    # ── 2. build alias map & snippets ──────────────────────────────────────
-    alias   = {}               #  e.g. {"baris": "Baris Firat", ...}
-    snippet_lines = []
-    for c in project["candidates"]:
-        fullname = c["name"]
-        first    = fullname.split()[0].lower()
-        alias[first] = fullname
-        snippet_lines.append(
-            f"- **{fullname}** (Fit {c['fit']} %): {c['text'][:250]}…"
-        )
-
-    alias_block = ", ".join(f"'{k}'→'{v}'" for k,v in alias.items())
-
-    # ── 3. compose system prompt ───────────────────────────────────────────
-    system_prompt = (
-        "You are an expert recruitment assistant.\n"
-        "Use ONLY the project description, markdown ranking table and résumé "
-        "snippets provided below. You are allowed to quote and compare the "
-        "candidates. Keep answers ≤120 words unless user asks for more.\n\n"
-        f"Alias map: {alias_block}\n\n"
-        "## Project description\n"
-        f"{project['description']}\n\n"
-        "## Ranking table (markdown)\n"
-        f"{project['table_md']}\n\n"
-        "## Candidate résumé snippets\n" +
-        "\n".join(snippet_lines)
-    )
-
-    # ── 4. add new user turn & send to GPT ─────────────────────────────────
-    history.append({"role":"user","content":user_text})
-    messages = [{"role":"system","content":system_prompt}] + history
-
-    response = openai.chat.completions.create(
-        model="gpt-3.5-turbo",
-        messages=messages,
-        temperature=0.4,
-        max_tokens=350
-    )
-    assistant_reply = response.choices[0].message.content.strip()
-
-    # ── 5. persist & return ────────────────────────────────────────────────
-    history.append({"role":"assistant","content":assistant_reply})
-    await chats.update_one(
-        {"user_id":user_id},
-        {"$set":{"messages":history}},
-        upsert=True
-    )
-    return {"reply": assistant_reply}
-
-
-    # ── Login ────────────────────────────────────────────────
-
-
-@app.get("/login", response_class=HTMLResponse)
-async def login_form(request: Request):
-       return await render(request, "login.html", page_title="Login", active="/login")
-
-@app.post("/login")
-async def login_post(request: Request,
-                     username: str = Form(...),
-                     password: str = Form(...)):
-    user = await verify_password(username, password)
-    if not user:
-        return await render(request, "login.html",
-                      {"error": "Invalid credentials"},
-                      page_title="Login", active="/login",
-                      status_code=401)
-    resp = RedirectResponse("/", status_code=303)
-    set_session(resp, user)
-    return resp
-
-@app.get("/logout")
-def logout():
-    resp = RedirectResponse("/", status_code=303)
-    clear_session(resp)
-    return resp
-
-
-
-@app.get("/admin/users", response_class=HTMLResponse, dependencies=[Depends(require_owner)])
-async def user_admin(request: Request):
-    cursor = users_coll.find({}, {"_id": 0, "hashed_password": 0})
-    users = await cursor.to_list(None)
-    return await render(request, "admin_users.html",
-                  {"users": users},
-                  page_title="User admin", active="/admin/users")
-
-@app.post("/admin/users", dependencies=[Depends(require_owner)])
-async def create_user_admin(
-    username: str = Form(...),
-    password: str = Form(...),
-    role: str = Form("user"),
-):
-    if await users_coll.find_one({"username": username}):
-        raise HTTPException(400, "User already exists")
-    await create_user(username, password, role)
-    return RedirectResponse("/admin/users", status_code=303)
-
-@app.post("/admin/users/delete", dependencies=[Depends(require_owner)])
-async def delete_user_admin(
-        username: str = Form(...),
-        current   = Depends(require_owner)):             # you are the owner
-    """
-    Deletes the account given by *username*.  
-    • Owner may not delete themselves to avoid lock-out.
-    """
-    if username == current["username"]:
-        raise HTTPException(400, "You cannot delete your own owner account")
-
-    res = await users_coll.delete_one({"username": username})
-    if res.deleted_count == 0:
-        raise HTTPException(404, "User not found")
-
-    return RedirectResponse("/admin/users", status_code=303)
-
-
-@app.get("/admin/users/{username}/edit", response_class=HTMLResponse,
-         dependencies=[Depends(require_owner)])
-async def edit_user_form(request: Request, username: str):
-    user = await users_coll.find_one({"username": username}, {"_id": 0, "hashed_password": 0})
-    if not user:
-        raise HTTPException(404, "User not found")
-    return await render(request, "edit_user.html",
-                  {"user": user},
-                  page_title="Edit user", active="/admin/users")
-
-
-@app.post("/admin/users/{username}/edit", dependencies=[Depends(require_owner)])
-async def update_user_admin(username: str, new_username: str = Form(...)):
-    if username != new_username:
-        if await users_coll.find_one({"username": new_username}):
-            raise HTTPException(400, "Username already exists")
-        res = await users_coll.update_one({"username": username}, {"$set": {"username": new_username}})
-        if res.matched_count == 0:
-            raise HTTPException(404, "User not found")
-    return RedirectResponse("/admin/users", status_code=303)
-
-
-@app.post("/admin/users/{username}/reset", dependencies=[Depends(require_owner)])
-async def reset_user_password(username: str):
-    res = await users_coll.update_one(
-        {"username": username},
-        {"$set": {"hashed_password": pwd_context.hash(DEFAULT_PASSWORD)}}
-    )
-    if res.matched_count == 0:
-        raise HTTPException(404, "User not found")
-    return RedirectResponse("/admin/users", status_code=303)
-
-
-@app.get("/profile", response_class=HTMLResponse, dependencies=[Depends(require_login)])
-async def profile(request: Request, user = Depends(get_current_user)):
-    return await render(request, "profile.html",
-                  {"user": user},
-                  page_title="Profile", active="/profile")
-
-# change password
-@app.post("/profile/password", dependencies=[Depends(require_login)])
-async def change_password(
-        old: str = Form(...),
-        new: str = Form(...),
-        user = Depends(get_current_user)
-    ):
-    if not await verify_password(user["username"], old):
-        raise HTTPException(400, "Old password incorrect")
-    await users_coll.update_one(
-        {"username": user["username"]},
-        {"$set": {"hashed_password": pwd_context.hash(new)}}
-    )
-    return RedirectResponse("/profile?ok=1", status_code=303)
+app.include_router(users_router)
+app.include_router(chat_router)
